+++ src/Command/Redis/XREADGROUP_CLAIM.php
@@ -32,6 +32,14 @@ class XREADGROUP_CLAIM extends RedisCommand
             array_push($processedArguments, 'COUNT', $arguments[3]);
         }
 
+        if (count($arguments) >= 8 && null !== $arguments[7]) {
+            array_push($processedArguments, 'MAXCOUNT', $arguments[7]);
+        }
+
+        if (count($arguments) >= 9 && null !== $arguments[8]) {
+            array_push($processedArguments, 'MAXSIZE', $arguments[8]);
+        }
+
         if (count($arguments) >= 5 && null !== $arguments[4]) {
             array_push($processedArguments, 'BLOCK', $arguments[4]);
         }
@@ -40,7 +48,7 @@ class XREADGROUP_CLAIM extends RedisCommand
             $processedArguments[] = 'NOACK';
         }
 
-        if (count($arguments) >= 7 && false !== $arguments[6]) {
+        if (count($arguments) >= 7 && false !== $arguments[6] && null !== $arguments[6]) {
             array_push($processedArguments, 'CLAIM', $arguments[6]);
         }