+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support for new TS commands + Indonesian language support integration test (#1695)
 - Added support for new COLLECT reducer for FT.AGGREGATE (#1699)
 - Added support for `SDIFFCARD` and `SUNIONCARD` command
+- Added support for `MAXCOUNT` and `MAXSIZE` arguments for `XREAD` and `XREADGROUP`
 - Added support for `TS.READ` command
 - Added support for `EXCLUDEEMPTY` argument for `TS.MRANGE` and `TS.MREVRANGE` commands
 - Added explicit testing for FT.SEARCH timeout policies